Enemies Made
Make an enemy as well as a "EnemyInput" script which will be similar to the "PlayerInput" script. Then using the "EnemyInput" and the "Movement" script to move the enemy.

To make the enemy walk back and forth I added a varible in the "EnemyInput" script that remembers where it spawns and another varible that manages how far it can go from its spawn before it needs to turn back

Two Current issues are 1. you need to set the distance for each enemy manually because I couldn't figure out how to automate it (yet) and 2. I still need to add the collisions for both the enemy and the player and how they will react